add gh autgenerated config
I have to add host.yml to gitignore because the gh-cli just stores your
auth token in plaintext?? In your .config too, which is pretty common to
sync to something like github. Seems kinda stupid to do this and not
even make it clear but what do I know. (I only became aware of this out because
github has a feature which tells you if your authtoken ends up in one of
your repos)